% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdownButtonInput.R
\name{dropdownButtonInput}
\alias{dropdownButtonInput}
\title{Dropdown button Shiny input widget}
\usage{
dropdownButtonInput(inputId, label, elements, width = NULL, class = NULL,
  icon = NULL, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The contents of the button or link–usually a text label, but
you could also use any other HTML, like an image.}

\item{elements}{A list of elements to display in the dropdown menu when the
button is clicked.  Each element in the list can be one of 3 things:  (1) a
named value, in which case the name will be displayed and the input bound
to value when it is clicked, (2) a shiny.tag object, which is displayed
as-is, or (3) NA which inserts a divider.}

\item{width}{Forced width for the dropdown menu.}

\item{class}{Additional element classes}

\item{icon}{An optional \code{\link{shiny::icon}} to appear on the
button.}

\item{...}{Additional attributes sent to \code{\link{shiny::div}}}
}
\value{
Shiny tag list suitable for use in a Shiny UI.
}
\description{
Dropdown button Shiny input widget
}
