% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdownButtonInput.R
\name{updateDropdownButtonInput}
\alias{updateDropdownButtonInput}
\title{Update the value of a dropdown button input}
\usage{
updateDropdownButtonInput(session, inputId, value = NULL)
}
\arguments{
\item{session}{Shiny server session object}

\item{inputId}{Input id for the dropdown button to update}

\item{value}{New value for the button}
}
\description{
Update the value of a dropdown button input
}
